+++ NEWS.md
@@ -1,3 +1,11 @@
+# felp 0.4.0
+
+- Support `fuzzyhelp` to run Shiny App in background without blocking user terminal.
+  The new behavior is enabled by default and can be disabled by passing `FALSE` to the `background` argument or to the `fuzzyhelp.background` option (#18, #20).
+- Fixed wrong behaviors of anchors in the HTML help in the UI of the `fuzzyhelp` function.
+  A click on a anchor should not cause nesting of the UI when href of the anchor is an ID.
+  Instead, the click should scroll the window to show the element with the corresponding ID (#17).
+
 # felp 0.3.0
 
 - Added `fuzzyhelp` function which launches Shiny Gadget to search help topics fuzzily, and preview the result. Done button will also launch `help` function.